# Docs/Api/cpay-v2-openapi.yaml
openapi: 3.0.3
info:
  title: CPay Gateway API
  version: 1.6.0
  description: Merchant payment, callback operations, reconciliation, finance, and operations endpoints for CPay.
security:
  - CPayV2Signature: []
paths:
...
  /api/v2/admin/reconciliation/settlements/close/approve:
    post:
      summary: Checker approval for a submitted settlement batch close
      operationId: approveSettlementBatchClose
      security:
        - AdminBasicAuth: []
      parameters:
        - name: reference
          in: query
          required: true
          schema:
            type: string
        - name: approvedBy
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Settlement batch closed
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v2/admin/reconciliation/settlements/close/reject:
    post:
      summary: Checker rejection for a submitted settlement batch close
      operationId: rejectSettlementBatchClose
      security:
        - AdminBasicAuth: []
      parameters:
        - name: reference
          in: query
          required: true
          schema:
            type: string
        - name: rejectedBy
          in: query
          required: true
          schema:
            type: string
        - name: reason
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Settlement batch marked rejected and left pending
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v2/admin/reconciliation/settlements/schedule:
    post:
      summary: Configure settlement sweep schedule
      operationId: scheduleSettlementSweep
      security:
        - AdminBasicAuth: []
      parameters:
        - name: merchantId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: provider
          in: query
          required: true
          schema:
            type: string
        - name: channel
          in: query
          required: true
          schema:
            type: string
        - name: currency
          in: query
          required: true
          schema:
            type: string
        - name: minimumRetainedBalance
          in: query
          required: false
          schema:
            type: string
        - name: sweepHour
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 23
      responses:
        '200':
          description: Settlement schedule saved
  /api/v2/admin/reconciliation/settlements/run-due:
    post:
      summary: Run due settlement sweeps
      operationId: runDueSettlementSweeps
      security:
        - AdminBasicAuth: []
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: hour
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Sweep results
...
components:
  securitySchemes:
    CPayV2Signature:
      type: apiKey
      in: header
      name: X-CPay-Signature
    AdminBasicAuth:
      type: http
      scheme: basic
...
  responses:
    BadRequest:
      description: Request was rejected
...